Upscaled some stst collection types to accomodate for the numbers present

+++ src/stats/construction_stats.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 namespace diplodocus {
 
 constexpr int kIntersectionCost = 2;
@@ -14,11 +16,11 @@ struct ConstructionStats {
     float prefix_scan_time = 0.0f;
     float merge_and_compact_time = 0.0f;
     float kernel_time = 0.0f;
-    int node_count = 0;
-    int inner_node_count = 0;
-    int leaf_node_count = 0;
+    int64_t node_count = 0;
+    int64_t inner_node_count = 0;
+    int64_t leaf_node_count = 0;
     float bvh_cost = 0.0f;
-    int memory_consumption = 0;
+    int64_t memory_consumption = 0;
 };
 
 }  // namespace diplodocus
